# api-docs.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "WordCounter API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/calculateMostFrequentNWords": {
      "post": {
        "tags": [
          "word-frequency-controller"
        ],
        "operationId": "calculateMostFrequentNWords",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WordFrequencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WordFrequency"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/calculateHighestFrequency": {
      "post": {
        "tags": [
          "word-frequency-controller"
        ],
        "operationId": "calculateHighestFrequency",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WordFrequencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WordFrequencyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/calculateFrequencyForWord": {
      "post": {
        "tags": [
          "word-frequency-controller"
        ],
        "operationId": "calculateFrequencyForWord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WordFrequencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WordFrequencyResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "WordFrequencyRequest": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "word": {
            "type": "string"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "WordFrequency": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string"
          },
          "frequency": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "WordFrequencyResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}