# src/main/java/com/ordina/countingwords/word/frequency/domain/WordFrequencyResponse.java
package com.ordina.countingwords.word.frequency.domain;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class WordFrequencyResponse {
    private String message;
}